implement proof construction
